add LoadingBackdrop component #3
