+++ src/components/LoadingBackdrop/index.tsx
@@ -1,0 +1,32 @@
+import {
+  Backdrop,
+  CircularProgress,
+  createStyles,
+  makeStyles,
+  Theme,
+} from '@material-ui/core';
+
+interface LoadingBackdrop {
+  isLoading: boolean;
+}
+
+const useStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    backdrop: {
+      zIndex: theme.zIndex.drawer + 1,
+      color: theme.palette.primary.dark,
+    },
+  })
+);
+
+const LoadingBackdrop: React.FC<LoadingBackdrop> = ({ isLoading }) => {
+  const classes = useStyles();
+
+  return (
+    <Backdrop className={classes.backdrop} open={isLoading}>
+      <CircularProgress color="inherit" />
+    </Backdrop>
+  );
+};
+
+export default LoadingBackdrop;
