+++ src/context/LoadingContext.tsx
@@ -1,11 +1,5 @@
-import {
-  Backdrop,
-  CircularProgress,
-  createStyles,
-  makeStyles,
-  Theme,
-} from '@material-ui/core';
 import { createContext, useCallback, useContext, useState } from 'react';
+import LoadingBackdrop from 'src/components/LoadingBackdrop';
 
 interface LoadingContextState {
   isLoading: boolean;
@@ -13,23 +7,12 @@ interface LoadingContextState {
   hideLoading(): void;
 }
 
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    backdrop: {
-      zIndex: theme.zIndex.drawer + 1,
-      color: theme.palette.primary.dark,
-    },
-  })
-);
-
 const LoadingContext = createContext<LoadingContextState>(
   {} as LoadingContextState
 );
 
 const LoadingProvider: React.FC = ({ children }) => {
   const [isLoading, setIsLoading] = useState(false);
-
-  const classes = useStyles();
 
   const showLoading = useCallback(() => {
     setIsLoading(true);
@@ -42,9 +25,7 @@ const LoadingProvider: React.FC = ({ children }) => {
   return (
     <LoadingContext.Provider value={{ isLoading, showLoading, hideLoading }}>
       {children}
-      <Backdrop className={classes.backdrop} open={isLoading}>
-        <CircularProgress color="inherit" />
-      </Backdrop>
+      <LoadingBackdrop isLoading={isLoading} />
     </LoadingContext.Provider>
   );
 };
